Refactor(#6) : 도메인 구조 리펙토링
