+++ group_service/settings.gradle
@@ -1,1 +1,6 @@
+pluginManagement {
+    plugins {
+        id 'org.jetbrains.kotlin.jvm' version '2.0.21'
+    }
+}
 rootProject.name = 'group_service'
