+++ group_service/src/main/java/io/urdego/group_service/api/service/GroupMemberService.java
@@ -1,0 +1,4 @@
+package io.urdego.group_service.api.service;
+
+public class GroupMemberServiceImpl {
+}
