--- group_service/src/main/java/io/urdego/group_service/application/port/in/GroupMemberService.java
@@ -1,4 +1,0 @@
-package io.urdego.group_service.application.port.in;
-
-public interface GroupMemberService {
-}
