--- group_service/src/main/java/io/urdego/group_service/application/service/GroupMemberServiceImpl.java
@@ -1,4 +1,0 @@
-package io.urdego.group_service.application.service;
-
-public class GroupMemberServiceImpl {
-}
