+++ group_service/src/main/java/io/urdego/group_service/domain/entity/groupMember/GroupMemberRole.java
@@ -1,0 +1,5 @@
+package io.urdego.group_service.domain.entity.groupMember;
+
+public enum GroupMemberRole {
+    MANAGER, MEMBER
+}
