+++ group_service/src/main/java/io/urdego/group_service/domain/entity/groupMember/repository/GroupMemberRepository.java
@@ -1,0 +1,8 @@
+package io.urdego.group_service.domain.entity.group.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface GroupMemberRepository extends JpaRepository<> {
+}
