--- group_service/src/main/java/io/urdego/group_service/infra/entity/group/GroupMemberRole.java
@@ -1,5 +1,0 @@
-package io.urdego.group_service.infra.entity.group;
-
-public enum GroupMemberRole {
-    MANAGER, MEMBER
-}
