--- group_service/src/main/java/io/urdego/group_service/infra/repository/GroupJPARepository.java
@@ -1,7 +1,0 @@
-package io.urdego.group_service.infra.repository;
-
-import io.urdego.group_service.infra.entity.group.GroupEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GroupJPARepository extends JpaRepository<GroupEntity, Long> {
-}
