--- group_service/src/main/java/io/urdego/group_service/infra/repository/GroupMemberJPARepository.java
@@ -1,7 +1,0 @@
-package io.urdego.group_service.infra.repository;
-
-import io.urdego.group_service.infra.entity.group.GroupMemberEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GroupMemberJPARepository extends JpaRepository<GroupMemberEntity, Long> {
-}
