--- group_service/src/main/java/io/urdego/group_service/infra/repository/GroupMemberRepositoryImpl.java
@@ -1,8 +1,0 @@
-package io.urdego.group_service.infra.repository;
-
-import io.urdego.group_service.application.port.out.GroupMemberRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class GroupMemberRepositoryImpl implements GroupMemberRepository {
-}
