--- group_service/src/main/java/io/urdego/group_service/infra/repository/GroupRepositoryImpl.java
@@ -1,13 +1,0 @@
-package io.urdego.group_service.infra.repository;
-
-import io.urdego.group_service.application.port.out.GroupRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class GroupRepositoryImpl implements GroupRepository {
-
-    private final GroupRepository groupRepository;
-
-
-
-}
